+++ bin/deploy_vpn.sh
@@ -1,0 +1,29 @@
+#!/usr/bin/env bash
+#
+# Ansible VPN deployment script. You can use it to deploy OpenVPN
+# either on your localhost machine or remote machine (address taken
+# from server config)
+#
+
+. $(dirname $0)/common.sh
+
+if [ -z "$1" ]; then
+    echo "Usage:"
+    echo "$0 [local | remote]"
+    echo ""
+    echo "local  - deploy OpenVPN server on the current machine (localhost)"
+    echo "remote - deploy OpenVPN server on remote machine (hostname from server configuration)"
+    exit 0
+fi
+
+
+if [ "${1}" = "local" ]; then
+    echo "Deploying on localhost"
+    ansible-playbook -i "${BIN_DIR}/local_inventory.sh" -K "${ANSIBLE_DIR}/local.yml"
+    exit $?
+fi
+
+if [ "${1}" = "remote" ]; then
+    ansible-playbook -i "${BIN_DIR}/inventory.sh" "${ANSIBLE_DIR}/remote.yml"
+    exit $?
+fi